TASK: Remove FormHelper and move all logic to Form, Field and Option Objects
