+++ Classes/Domain/Model/AbstractFormHelper.php
@@ -1,0 +1,106 @@
+<?php
+declare(strict_types=1);
+
+namespace Neos\Fusion\Form\Domain\Model;
+
+use Neos\Eel\ProtectedContextAwareInterface;
+
+abstract class AbstractFormHelper implements ProtectedContextAwareInterface
+{
+
+    /**
+     * Convert a value to a string representation for beeing rendered as an html form value
+     *
+     * @param mixed $value
+     * @return string|null
+     */
+    protected function stringifyValue($value): string
+    {
+        if (is_object($value)) {
+            $identifier = $this->persistenceManager->getIdentifierByObject($value);
+            if ($identifier !== null) {
+                return $identifier;
+            }
+        }
+        return (string)$value;
+    }
+
+    /**
+     * Convert an array of values to an array of string representation for beeing rendered as an html form value
+     *
+     * @param array $value
+     * @return array|null
+     */
+    protected function stringifyArray(array $value = null): array
+    {
+        if (is_array($value)) {
+            return array_map(
+                function ($value) {
+                    return $this->stringifyValue($value);
+                },
+                $value
+            );
+        } else {
+            return null;
+        }
+    }
+
+
+    /**
+     * Prepend the given fieldNamePrefix to the fieldName the
+     *
+     * @param string $name
+     * @param string|null $prefix
+     * @return string
+     */
+    protected function prefixFieldName(string $fieldName, string $fieldNamePrefix = null): string
+    {
+        if (!$fieldNamePrefix) {
+            return $fieldName;
+        } else {
+            $fieldNameSegments = explode('[', $fieldName, 2);
+            $fieldName = $fieldNamePrefix . '[' . $fieldNameSegments[0] . ']';
+            if (count($fieldNameSegments) > 1) {
+                $fieldName .= '[' . $fieldNameSegments[1];
+            }
+            return $fieldName;
+        }
+    }
+
+    /**
+     * Convert the given html fieldName to a dot seperated path
+     *
+     * @param $name
+     * @return string
+     */
+    protected function fieldNameToPath($name): string
+    {
+        $path = preg_replace('/(\]\[|\[|\])/', '.', $name);
+        return trim($path, '.');
+    }
+
+    /**
+     * Convert the given dot seperated $path to an html fieldName
+     *
+     * @param $name
+     * @return string
+     */
+    protected function pathToFieldName($path): string
+    {
+        $pathSegments = explode('.', $path);
+        $fieldName = array_shift($pathSegments);
+        foreach ($pathSegments as $pathSegment) {
+            $fieldName .= '[' . $pathSegment . ']';
+        }
+        return $fieldName;
+    }
+
+    /**
+     * @param string $methodName
+     * @return bool
+     */
+    public function allowsCallOfMethod($methodName)
+    {
+        return true;
+    }
+}
