+++ Classes/FusionObjects/FieldDefinition.php
@@ -13,21 +13,14 @@ namespace Neos\Fusion\Form\FusionObjects;
  * source code.
  */
 
-use Neos\Flow\Annotations as Flow;
 use Neos\Fusion\FusionObjects\AbstractFusionObject;
 use Neos\Fusion\Form\Domain\Model\Form;
 use Neos\Fusion\Form\Domain\Model\Field;
-use Neos\Fusion\Form\Eel\FormHelper;
 use Neos\Error\Messages\Result;
 use Neos\Utility\ObjectAccess;
 
 class FieldDefinition extends AbstractFusionObject
 {
-    /**
-     * @var FormHelper
-     * @Flow\Inject
-     */
-    protected $formHelper;
 
     /**
      * @return Field
@@ -40,50 +33,16 @@ class FieldDefinition extends AbstractFusionObject
         $value = $this->fusionValue('value');
         $multiple = $this->fusionValue('multiple');
 
-        // early return
-        if (!$name) {
-            if ($outerField && $outerField instanceof Field) {
-                return $outerField->withTargetValue($value);
-            }
-            return new Field(null, null, null, $value);
-        }
-
-        // render fieldName
-        if ($form && $form->getFieldNamePrefix()) {
-            $fieldName = $this->formHelper->prefixFieldName($name, $form->getFieldNamePrefix());
-        } else {
-            $fieldName = $name;
-        }
-        if ($multiple) {
-            $fieldName .= '[]';
-        }
-
-        // create property path from fieldname
-        $path = $this->formHelper->fieldNameToPath($name);
-
-        // determine value, according to the following algorithm:
-        if ($form && $form->getResult() !== null && $form->getResult()->hasErrors()) {
-            // 1) if a validation error has occurred, pull the value from the submitted form values.
-            $fieldValue = ObjectAccess::getPropertyPath($form->getSubmittedValues(), $path);
-        } elseif ($path && $form && $form->getData()) {
-            // 2) else, if "property" is specified, take the value from the bound object.
-            $fieldValue = ObjectAccess::getPropertyPath($form->getData(), $path);
-        } else {
-            $fieldValue = null;
-        }
-
-        // determine ValidationResult for the single property
-        $fieldResult = null;
-        if ($form && $form->getResult() && $form->getResult()->hasErrors()) {
-            $fieldResult = $form->getResult()->forProperty($path);
+        // reuse outerfield if no name is given
+        if (!$name && $outerField && $outerField instanceof Field) {
+            return $outerField->withTargetValue($value);
         }
 
         return new Field (
-            $fieldName,
-            $fieldValue,
+            $form,
+            $name,
             $value,
-            $multiple,
-            $fieldResult
+            $multiple
         );
     }
 }
