+++ Configuration/Settings.yaml
@@ -1,8 +1,4 @@
 Neos:
-  Fusion:
-    defaultContext:
-      Form: 'Neos\Fusion\Form\Eel\FormHelper'
-
   Neos:
     fusion:
       autoInclude:
