+++ Resources/Private/Fusion/Prototypes/Button.fusion
@@ -5,7 +5,7 @@ prototype(Neos.Fusion.Form:Button) < prototype(Neos.Fusion.Form:FieldComponent) 
     renderer = afx`
         <button
             name={field.getName()}
-            value={Form.stringifyValue(field.getValue())}
+            value={field.getTargetValue(true)}
             {...props.attributes}
         >
             {props.content}
