+++ Resources/Private/Fusion/Prototypes/Input.fusion
@@ -9,7 +9,7 @@ prototype(Neos.Fusion.Form:Input)  < prototype(Neos.Fusion.Form:FieldComponent) 
     renderer = afx`
         <input
             name={field.getName()}
-            value={Form.stringifyValue(field.getValue() || field.getTargetValue())}
+            value={field.getCurrentValue(true) || field.getTargetValue(true)}
             {...props.attributes}
         />
     `
