+++ Resources/Private/Fusion/Prototypes/Select.Option.fusion
@@ -43,20 +43,20 @@ prototype(Neos.Fusion.Form:Select.Option)  < prototype(Neos.Fusion:Component) {
         # the `selected` state is calculated outside the renderer to allow` overriding via `attributes`
         #
         selected = false
-        selected.@process.checkMultiValue = ${Array.indexOf(Form.stringifyArray(field.getValue()), Form.stringifyValue(option.getValue())) > -1}
-        selected.@process.checkMultiValue.@if.hasValue = ${field.getValue()}
+        selected.@process.checkMultiValue = ${Array.indexOf(field.getCurrentValue(true), option.getTargetvalue(true)) > -1}
+        selected.@process.checkMultiValue.@if.hasValue = ${field.hasCurrentValue()}
         selected.@process.checkMultiValue.@if.isMultiple = ${field.isMultiple()}
-        selected.@process.checkSingleValue = ${Form.stringifyValue(field.getValue()) == Form.stringifyValue(option.getValue())}
-        selected.@process.checkSingleValue.@if.hasValue = ${field.getValue()}
+        selected.@process.checkSingleValue = ${field.getCurrentValue(true) == option.getTargetvalue(true)}
+        selected.@process.checkSingleValue.@if.hasValue = ${field.hasCurrentValue()}
         selected.@process.checkSingleValue.@if.isSingle = ${!field.isMultiple()}
 
         renderer = afx`
            <option
-               value={Form.stringifyValue(option.getValue())}
+               value={option.getTargetValue(true)}
                selected={props.selected}
                {...props.attributes}
            >
-               {props.content || option.getValue()}
+               {props.content || option.getTargetValue()}
            </option>
         `
     }
